Create _P118_MPU6050.ino
This plugin lets the user read the raw values of the sensor, supports both addresses and can be set to monitor movement by setting range-threshold values for each axis (or disable monitoring for a axis) and set a movement detection threshold.

By first monitoring the raw values, a baseline can be determined using the home controllers ability to graph the values. The thresholds can then be added to the plugin and a timeframe can be set in which the thresholds need to be exceeded a programmable number of times before movement is singled as a switch. By adding multiple instances of the plugin, multiple movement detection schemes can be set, this way it is possible to track a (e.g. washing) cycle.