Updated Thingspeak controller plugin
This update addresses some of the limitations of the Thingspeak server, such as only 8 fields are allowed to updated and period between 2 update requests cannot be less than 15s.

Regardless of what the sensor delays are set to, every time the plugin is called it will store the most recent values from the sensor in a buffer (only for values corresponding to 8 fields in Thingspeak), and check if 15s have elapsed since last publish.
If 15s have elapsed since last publish, it will check if there is any data available in the buffer and publish it.
If no data is available in buffer then publish request is not sent.

This update eliminates the problem with multiple sensors publishing in 15s time period, resulting in only one publish being successful and others ignored. With this update, any new sensor values received within 15s of last publish will be stored in buffer and will be sent when next time the plugin is called and 15s have elapsed.