Add file P202_ADC_ACcurrentSensor.ino
Add file: plugin _P202 to read AC current.

This is a work-in-progress. The plugin reads the mains (line) AC current via a current transformer (C.T.), currently a TA12-100 C.T. sensor, with potential to be adapted for other C.T. based sensors. 
This Plugin is an extension of the existing ADC plugin (_P002) and incorporates code from a Henry's Bench Arduino sketch to identify and record the peak current flow during the 50 or 60 Hz altenating current cyce. The Plugin also contains a number of parameters which can be configured at run-time. The value of AC current is also converted to Watts using a typical (configurable) value for AC mains (line) volts at the user's location.