extend MQTT Topic
I use MQTT for my Home Assistant setup and I am very happy with it. Now I figured out that the max length is just 40 which is not enough as I need more for my different topics. Is it with this Pull request possible having a longer Topic? I Use the binariy and the flashtool and dunno how I would change the topic size to 80